organise skibidi into more packages and add a stylecheck
