+++ build.gradle
@@ -1,7 +1,12 @@
 plugins {
     id 'java'
     id 'application'
+    id 'checkstyle'
     id 'com.github.johnrengelman.shadow' version '7.1.2'
+}
+
+checkstyle {
+    toolVersion = '10.2'
 }
 
 repositories {
@@ -28,7 +33,7 @@ test {
 }
 
 application {
-    mainClass.set("skibidi.Skibidi")
+    mainClass.set("skibidi.main.Skibidi")
 }
 
 shadowJar {
