+++ src/main/java/skibidi/command/InvalidCommand.java
@@ -1,5 +1,12 @@
-package skibidi;
+package skibidi.command;
 
+import skibidi.utils.Storage;
+import skibidi.utils.TaskList;
+import skibidi.utils.Ui;
+
+/**
+ * Represents an invalid command.
+ */
 public class InvalidCommand extends Command {
     public static final String COMMAND_WORD = "invalid";
