+++ src/main/java/skibidi/command/ListCommand.java
@@ -1,5 +1,12 @@
-package skibidi;
+package skibidi.command;
 
+import skibidi.utils.Storage;
+import skibidi.utils.TaskList;
+import skibidi.utils.Ui;
+
+/**
+ * Represents a command to list all tasks in the task list.
+ */
 public class ListCommand extends Command {
 
     public static final String COMMAND_WORD = "list";
