+++ src/main/java/skibidi/exceptions/SkibidiException.java
@@ -1,10 +1,10 @@
-package skibidi;
+package skibidi.exceptions;
 
 /**
  * Represents an exception that is thrown when an error occurs in Skibidi.
  */
 public class SkibidiException extends Exception {
-    public SkibidiException (String message) {
+    public SkibidiException(String message) {
         super(message);
     }
 }
