+++ src/main/java/skibidi/task/Deadline.java
@@ -1,4 +1,6 @@
-package skibidi;
+package skibidi.task;
+
+import skibidi.utils.MilitaryTime;
 
 /**
  * Represents a deadline task.
@@ -7,6 +9,13 @@ public class Deadline extends Task {
     protected String by;
     private MilitaryTime time;
 
+    /**
+     * Creates a deadline task.
+     *
+     * @param description The description of the deadline task.
+     * @param by The deadline of the deadline task.
+     * @param time The time of the deadline task.
+     */
     public Deadline(String description, String by, MilitaryTime time) {
         super(description);
         this.by = by;
