+++ src/main/java/skibidi/task/Event.java
@@ -1,4 +1,6 @@
-package skibidi;
+package skibidi.task;
+
+import skibidi.utils.MilitaryTime;
 
 /**
  * Represents an event task.
@@ -9,6 +11,15 @@ public class Event extends Task {
     private MilitaryTime startTime;
     private MilitaryTime endTime;
 
+    /**
+     * Creates an event task.
+     *
+     * @param description The description of the event task.
+     * @param from The starting location of the event task.
+     * @param startTime The starting time of the event task.
+     * @param to The ending location of the event task.
+     * @param endTime The ending time of the event task.
+     */
     public Event(String description, String from, MilitaryTime startTime, String to, MilitaryTime endTime) {
         super(description);
         this.from = from;
@@ -19,8 +30,8 @@ public class Event extends Task {
 
     @Override
     public String toString() {
-        return " [E]" + super.toString() + " (from: " + from + " " + startTime.toString() +
-                " to: " + to + " " + endTime.toString() + ")";
+        return " [E]" + super.toString() + " (from: " + from + " " + startTime.toString()
+                + " to: " + to + " " + endTime.toString() + ")";
     }
 
     @Override
