+++ src/main/java/skibidi/task/Todo.java
@@ -1,10 +1,10 @@
-package skibidi;
+package skibidi.task;
 
 /**
  * Represents a todo task.
  */
 public class Todo extends Task {
-    public Todo (String description) {
+    public Todo(String description) {
         super(description);
     }
