+++ src/main/java/skibidi/utils/MilitaryTime.java
@@ -1,17 +1,30 @@
-package skibidi;
+package skibidi.utils;
+
+import skibidi.exceptions.SkibidiException;
 
 /**
  * Represents a time in 24-hour format.
  */
 public class MilitaryTime {
     protected String time;
+    /**
+     * Creates a time in 24-hour format.
+     *
+     * @param time The time in 24-hour format.
+     * @throws SkibidiException If the time is not in the correct format.
+     */
     public MilitaryTime(String time) throws SkibidiException {
         if (!isValidTime(time)) {
             throw new SkibidiException("Invalid time format. Please enter time in HHmm format.");
         }
         this.time = time;
     }
-
+    /**
+     * Checks if the time is in the correct format.
+     *
+     * @param time The time to be checked.
+     * @return True if the time is in the correct format, false otherwise.
+     */
     public boolean isValidTime(String time) {
         if (time.length() != 4) {
             return false;
