+++ src/main/java/skibidi/utils/TimeUtil.java
@@ -1,9 +1,15 @@
-package skibidi;
+package skibidi.utils;
 
 /**
  * Represents a utility class for converting time between standard and military time.
  */
 public class TimeUtil {
+    /**
+     * Converts standard time to military time.
+     *
+     * @param standardTime Standard time to be converted.
+     * @return Standard time.
+     */
     public static String convertStandardToMilitary(String standardTime) {
         String[] parts = standardTime.split(" ");
         String time = parts[0];
