+++ src/test/java/skibidi/task/DeadlineTest.java
@@ -1,7 +1,12 @@
-package skibidi;
+package skibidi.task;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import skibidi.exceptions.SkibidiException;
+import skibidi.utils.MilitaryTime;
+
 
 public class DeadlineTest {
