+++ src/test/java/skibidi/task/TodoTest.java
@@ -1,7 +1,9 @@
-package skibidi;
+package skibidi.task;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.assertEquals;
+
 
 public class TodoTest {
